+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   ([#7740](https://github.com/mitmproxy/mitmproxy/pull/7740), @mhils)
 - Fix crash in mitmweb when no explicit Server-Connection is logged.
   ([#7734](https://github.com/mitmproxy/mitmproxy/pull/7734), @lups2000)
+- Add syntax highlighting for CSS and JavaScript contentviews.
 
 ## 25 May 2025: mitmproxy 12.1.1